lf_inspect.py : use float when there is a possibility of the number being NaN

+++ py-scripts/tools/lf_inspect.py
@@ -19,6 +19,7 @@ import logging
 import re
 import csv
 import traceback
+import math
 
 
 sys.path.append(os.path.join(os.path.abspath(__file__ + "../../../../")))
@@ -309,8 +310,8 @@ class inspect_sql:
                             logger.debug("type: {data} {data2}".format(data=type(df_data_2), data2=df_data_2))
 
                             percent_delta = 0
-                            if((int(df_data_1['numeric-score']) != 0 and df_data_1['numeric-score'] is not None) and df_data_2 is not None):
-                                percent_delta = round(((df_data_2['numeric-score']/df_data_1['numeric-score']) * 100), 2)
+                            if((float(df_data_1['numeric-score']) != 0 and df_data_1['numeric-score'] is not None) and df_data_2 is not None):
+                                percent_delta = round(((float(df_data_2['numeric-score'])/float(df_data_1['numeric-score'])) * 100), 2)
 
                             if percent_delta >= 90:
                                 logger.info("Performance Good {percent}".format(percent=percent_delta))
@@ -327,7 +328,7 @@ class inspect_sql:
                             elif percent_delta == 0:
                                 # negative logic like Stations Failed IP if zero is a goot thing
                                 if 'Failed' in df_data_1['short-description']:
-                                    if((int(df_data_1['numeric-score']) != 0) or (int(df_data_2['numeric-score']) !=0)):
+                                    if((float(df_data_1['numeric-score']) != 0) or (float(df_data_2['numeric-score']) !=0)):
                                         logger.info("Performance Critical {percent}".format(percent=percent_delta))
                                         background = self.background_red
                                         self.test_result = "Critical"
@@ -511,7 +512,7 @@ class inspect_sql:
                         df_data_2 = df_tmp.iloc[db_index_2]
 
                         percent_delta = 0
-                        if((int(df_data_1['numeric-score']) != 0 and df_data_1['numeric-score'] is not None) and df_data_2 is not None):
+                        if((float(df_data_1['numeric-score']) != 0 and df_data_1['numeric-score'] is not None) and df_data_2 is not None):
                             percent_delta = round(((df_data_2['numeric-score']/df_data_1['numeric-score']) * 100), 2)
 
                         if percent_delta >= 90:
@@ -529,7 +530,7 @@ class inspect_sql:
                         elif percent_delta == 0:
                             # negative logic like Stations Failed IP if zero is a goot thing
                             if 'Failed' in df_data_1['short-description']:
-                                if((int(df_data_1['numeric-score']) != 0) or (int(df_data_2['numeric-score']) !=0)):
+                                if((float(df_data_1['numeric-score']) != 0) or (float(df_data_2['numeric-score']) !=0)):
                                     logger.info("Performance Critical {percent}".format(percent=percent_delta))
                                     background = self.background_red
                                     self.test_result = "Critical"
@@ -770,8 +771,8 @@ class inspect_sql:
                             logger.debug("type: {data} {data2}".format(data=type(df_data_2), data2=df_data_2))
 
                             percent_delta = 0
-                            if((int(df_data_1['numeric-score']) != 0 and df_data_1['numeric-score'] is not None) and df_data_2 is not None):
-                                percent_delta = round(((df_data_2['numeric-score']/df_data_1['numeric-score']) * 100), 2)
+                            if((float(df_data_1['numeric-score']) != 0 and df_data_1['numeric-score'] is not None) and df_data_2 is not None):
+                                percent_delta = round(((float(df_data_2['numeric-score'])/float(df_data_1['numeric-score'])) * 100), 2)
 
                             if percent_delta >= 90:
                                 logger.info("Performance Good {percent}".format(percent=percent_delta))
@@ -788,7 +789,7 @@ class inspect_sql:
                             elif percent_delta == 0:
                                 # negative logic like Stations Failed IP if zero is a goot thing
                                 if 'Failed' in df_data_1['short-description']:
-                                    if((int(df_data_1['numeric-score']) != 0) or (int(df_data_2['numeric-score']) !=0)):
+                                    if((float(df_data_1['numeric-score']) != 0) or (float(df_data_2['numeric-score']) !=0)):
                                         logger.info("Performance Critical {percent}".format(percent=percent_delta))
                                         background = self.background_red
                                         self.test_result = "Critical"
